Extract the parse method into utils
